+++ src/helpers.php
@@ -1,0 +1,8 @@
+<?php
+
+if (! function_exists('llt')) {
+    function llt(): LaraLogsToolkit
+    {
+        return app(LaraLogsToolkit::class);
+    }
+}